Lighting 2.0 and Post Proccesing
This is a further review of lighting and post processing. I did my work on a piece of the level design created by James to explore the effects. I started working on lightmapping, and testing the efficiency of bakes based on the complexity of the map. Ran into a few light issues along the way. Still want clearer shadows, and more control over the final look. Learned to build skyboxes, and change the overall tone. Made a glass shader and material in blender only to realize that doesnt transfer over, so going to try and repeat that process in unity.