Negative Space Version 1.0
This is the first combination of concepts created into one location. Files, assets, and packages have been organized along with aspects from each persons work.

This involves a main level scene which contains lighting elements, player movement, the implementation of object rotation, a folder of collected sounds, and the beginning of a flashlight. This has all been tweaked to work together in our level design.

Currently the player can walk in a very small area around the center of the level.