Implemented First Puzzle
First Puzzle in place. Rotator works on a parent object which has a correct rotation of 0,0,0. Camera switch feature is not working for some reason. Current Control script is crashing the game, so an over simplified one is currently in use just so I can test objects in the scene.