Fixed Inner Dome issue
The issue was caused by having only one mesh as the dome and the inner face was not rendered.
I added a second Dome with the Normals Flipped and placed it int the scene and animated it.
The dome still needs lighting techniques applied so it wont look like a boxy dome